refactor(settings): move the deep-link section union to its own module
Adding a settings destination now touches one small module instead of
the chat store types.
