+++ src/renderer/src/store/chat-store-types.ts
@@ -215,32 +215,8 @@ export type ClearDesignHistoryResult = {
 }
 
 export type InitialSetupMode = 'required' | 'preview'
-export type SettingsRouteSection =
-  | 'general'
-  | 'providers'
-  | 'extensions'
-  | 'write'
-  | 'design'
-  | 'imageGeneration'
-  | 'mediaGeneration'
-  | 'speechToText'
-  | 'agents'
-  | 'laboratory'
-  | 'subagents'
-  | 'archives'
-  | 'worktree'
-  | 'memory'
-  | 'permissions'
-  | 'skill'
-  | 'mcp'
-  | 'shortcuts'
-  | 'easterEgg'
-  | 'claw'
-  | 'updates'
-  | 'terminal'
-  | 'debug'
-  | 'storage'
-  | 'dataMigration'
+import type { SettingsRouteSection } from './settings-route-sections'
+export type { SettingsRouteSection }
 export type AppRoute = 'chat' | 'write' | 'design' | 'settings' | 'plugins' | 'extensions' | 'claw' | 'board' | 'schedule' | 'workflow'
 export type ThreadCompletionOutcome = 'completed' | 'failed'
 export type CompletionAttentionRegistry = Record<string, ThreadCompletionOutcome | boolean>
